feedback: move exported Recommend context helper function
